# backend/openapi.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: current
  description: The API for current
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/CSCC01-current/currentapi/1.0.0
  - url: http://localhost:4200
paths:
...
  /api/v1/groups/{_id}/calls:
    parameters:
      - name: _id
        description: The mongo _id of a group
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/MongoId"
    get:
      summary: Get calls in a group by _id. Must be signed in and a member of the group.
      responses:
        200:
          description: The calls in the group corresponding to the provided _id.
          content:
            application/json:
              schema:
                type: object
                required:
                  - calls
                properties:
                  calls:
                    type: array
                    items:
                      $ref: "#/components/schemas/Call"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/InternalServerError"
    post:
      summary: Create a call in a group. Must be signed in and a member of the group.
      # validating form data not working
      # requestBody:
      #   required: true
      #   content:
      #     multipart/form-data:
      #       schema:
      #         type: object
      #         properties:
      #           audioFile:
      #             type: string
      #             format: binary
      responses:
        201:
          description: Successfully created the call.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        400:
          $ref: "#/components/responses/BadRequest"
        403:
          $ref: "#/components/responses/Forbidden"
        409:
          $ref: "#/components/responses/Conflict"
        500:
          $ref: "#/components/responses/InternalServerError"
...
components:
  schemas:
...
    MongoId:
      type: string
      # some format?
    Call:
      type: object
      required:
        - _id
        - audioFile
        - startTimestamp
        - endTimestamp
        - transcript
      properties:
        _id:
          readOnly: true
          allOf:
            - $ref: "#/components/schemas/MongoId"
        audioFile:
          type: string
          format: uri
        startTimestamp:
          type: string
          format: date-time
        endTimestamp:
          type: string
          format: date-time
        transcript:
          type: string
    Errors:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: "#/components/schemas/Error"
    Error:
      type: object
      required:
        - path
        - message
        - detail
      properties:
        path:
          type: string
        message:
          type: string
        detail:
          type: string
  responses:
    BadRequest:
      description: request format is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"
...
    Forbidden:
      description: insufficient permission for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"
    NotFound:
      description: specified resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"
    Conflict:
      description: request has correct format but data currently conflicts
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"
    InternalServerError:
      description: error processing request on server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Errors"